Fix visual offsets

+++ src/AvaloniaEdit/Text/TrailingInfo.cs
@@ -3,6 +3,6 @@ namespace AvaloniaEdit.Text
     internal sealed class TrailingInfo
     {
         public int Count { get; set; }
-        public int SpaceWidth { get; set; }
+        public double SpaceWidth { get; set; }
     }
 }
